Refactor Move class into subclasses (#42)
Do the refactor into separate subclasses.
